# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Marketplace API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface MessageResponse {
  message: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  customer: 'customer',
  seller: 'seller',
} as const;

export interface User {
  id: number;
  email: string;
  name: string;
  role: UserRole;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export type RegisterInputRole = typeof RegisterInputRole[keyof typeof RegisterInputRole];


export const RegisterInputRole = {
  customer: 'customer',
  seller: 'seller',
} as const;

export interface RegisterInput {
  email: string;
  password: string;
  name: string;
  role: RegisterInputRole;
}

export type LoginInputRole = typeof LoginInputRole[keyof typeof LoginInputRole];


export const LoginInputRole = {
  customer: 'customer',
  seller: 'seller',
} as const;

export interface LoginInput {
  email: string;
  password: string;
  role: LoginInputRole;
}

export interface Product {
  id: number;
  sellerId: number;
  sellerName: string;
  name: string;
  description: string;
  price: number;
  /** @nullable */
  discountPercent: number | null;
  finalPrice: number;
  category: string;
  stock: number;
  /** @nullable */
  imageUrl: string | null;
  createdAt: string;
}

export interface ProductInput {
  name: string;
  description: string;
  price: number;
  category: string;
  stock: number;
  /** @nullable */
  imageUrl?: string | null;
}

export interface ProductUpdate {
  name?: string;
  description?: string;
  price?: number;
  category?: string;
  /** @nullable */
  imageUrl?: string | null;
}

export interface DiscountInput {
  /** @nullable */
  discountPercent: number | null;
}

export interface StockUpdate {
  stock: number;
}

export interface CartItem {
  productId: number;
  product: Product;
  quantity: number;
  subtotal: number;
}

export interface Cart {
  items: CartItem[];
  total: number;
  itemCount: number;
}

export interface CartItemInput {
  productId: number;
  quantity: number;
}

export interface CartItemUpdate {
  quantity: number;
}

export interface OrderItem {
  productId: number;
  productName: string;
  quantity: number;
  unitPrice: number;
  subtotal: number;
}

export type OrderStatus = typeof OrderStatus[keyof typeof OrderStatus];


export const OrderStatus = {
  pending: 'pending',
  processing: 'processing',
  shipped: 'shipped',
  delivered: 'delivered',
  cancelled: 'cancelled',
} as const;

export interface Order {
  id: number;
  customerId: number;
  customerName: string;
  customerEmail: string;
  items: OrderItem[];
  total: number;
  status: OrderStatus;
  shippingAddress: string;
  createdAt: string;
  updatedAt: string;
}

export interface OrderInput {
  shippingAddress: string;
}

export type OrderStatusUpdateStatus = typeof OrderStatusUpdateStatus[keyof typeof OrderStatusUpdateStatus];


export const OrderStatusUpdateStatus = {
  pending: 'pending',
  processing: 'processing',
  shipped: 'shipped',
  delivered: 'delivered',
  cancelled: 'cancelled',
} as const;

export interface OrderStatusUpdate {
  status: OrderStatusUpdateStatus;
}

export interface StatusCount {
  status: string;
  count: number;
}

export interface SellerDashboard {
  totalProducts: number;
  totalOrders: number;
  totalRevenue: number;
  pendingOrders: number;
  lowStockProducts: number;
  recentOrders: Order[];
  ordersByStatus: StatusCount[];
}

export interface CustomerDashboard {
  totalOrders: number;
  pendingOrders: number;
  deliveredOrders: number;
  totalSpent: number;
  recentOrders: Order[];
}

export type ListProductsParams = {
/**
 * @nullable
 */
category?: string | null;
/**
 * @nullable
 */
search?: string | null;
/**
 * @nullable
 */
sellerId?: number | null;
};