# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized

  # PRODUCTS
...
  /cart:
    get:
      operationId: getCart
      tags: [cart]
      summary: Get current user's cart
      responses:
        "200":
          description: Cart with items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
        "401":
          description: Unauthorized
  /cart/items:
    post:
      operationId: addToCart
      tags: [cart]
      summary: Add item to cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CartItemInput"
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
  /cart/items/{productId}:
    patch:
      operationId: updateCartItem
      tags: [cart]
      summary: Update cart item quantity
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CartItemUpdate"
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"

    delete:
      operationId: removeFromCart
      tags: [cart]
      summary: Remove item from cart
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders (customer sees own; seller sees incoming)
      responses:
        "200":
          description: Orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"

    post:
      operationId: placeOrder
      tags: [orders]
      summary: Place an order from cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Order placed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Not found
  /orders/{id}/status:
    patch:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status (seller only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderStatusUpdate"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  # DASHBOARD SUMMARIES
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [customer, seller]
        createdAt:
          type: string
      required: [id, email, name, role, createdAt]
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required: [user, token]
    RegisterInput:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [customer, seller]
      required: [email, password, name, role]
    LoginInput:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [customer, seller]
      required: [email, password, role]
    Product:
      type: object
      properties:
        id:
          type: integer
        sellerId:
          type: integer
        sellerName:
          type: string
        name:
          type: string
        description:
          type: string
        price:
          type: number
        discountPercent:
          type: ["number", "null"]
        finalPrice:
          type: number
        category:
          type: string
        stock:
          type: integer
        imageUrl:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, sellerId, sellerName, name, description, price, discountPercent, finalPrice, category, stock, imageUrl, createdAt]
...
    CartItem:
      type: object
      properties:
        productId:
          type: integer
        product:
          $ref: "#/components/schemas/Product"
        quantity:
          type: integer
        subtotal:
          type: number
      required: [productId, product, quantity, subtotal]
    Cart:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItem"
        total:
          type: number
        itemCount:
          type: integer
      required: [items, total, itemCount]
    CartItemInput:
      type: object
      properties:
        productId:
          type: integer
        quantity:
          type: integer
      required: [productId, quantity]
    CartItemUpdate:
      type: object
      properties:
        quantity:
          type: integer
      required: [quantity]
    OrderItem:
      type: object
      properties:
        productId:
          type: integer
        productName:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: number
        subtotal:
          type: number
      required: [productId, productName, quantity, unitPrice, subtotal]
    Order:
      type: object
      properties:
        id:
          type: integer
        customerId:
          type: integer
        customerName:
          type: string
        customerEmail:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        total:
          type: number
        status:
          type: string
          enum: [pending, processing, shipped, delivered, cancelled]
        shippingAddress:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, customerId, customerName, customerEmail, items, total, status, shippingAddress, createdAt, updatedAt]
    OrderInput:
      type: object
      properties:
        shippingAddress:
          type: string
      required: [shippingAddress]
    OrderStatusUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [pending, processing, shipped, delivered, cancelled]
      required: [status]
...