# lib/api-zod/src/generated/types/loginInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Marketplace API
 * OpenAPI spec version: 0.1.0
 */
import type { LoginInputRole } from './loginInputRole';

export interface LoginInput {
  email: string;
  password: string;
  role: LoginInputRole;
}