# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: number;
  sellerId: number;
  sellerName: string;
  name: string;
  description: string;
  price: number;
  /** @nullable */
  discountPercent: number | null;
  finalPrice: number;
  category: string;
  stock: number;
  /** @nullable */
  imageUrl: string | null;
  createdAt: string;
}